Added bazel intellij plugin
There is a nice infrastructure in Intellij that allows smooth and easy
integration of external build systems, bazel in this case.

There are certain things that need to be implemented in the future for a better
experience, like syntax highlighting for the BUILD files and their integration/inclusion
into intellij infrastructure.

Signed-off-by: Nikita Vakula <programmistov.programmist@gmail.com>